oce: Kill the variant linking script
The only variant-specific RIL blob is now the QC libril,
which can simply be shipped among the others and selected
during runtime based on the rild.libpath system property.

The motivation for doing this after all these years is that,
as of Q, with the enablement of system-as-root it was not
easy to adapt this script and make it recovery-agnostic
at the same time.

[TARKZiM] Drop entire releasetools.

Change-Id: Ie863455f66a9561f405b6a9fa5b1ce0cb71ae432